+++ CHANGELOG.md
@@ -17,6 +17,9 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 - Added endpoint `PUT /provider` as an idempotent way of creating a
   provider. (#28)
 
+- Add endpoint `PUT /token` as an idempotent way of creating a
+  token. (#26)
+
 - Added environment var for setting bind address and port. (#29)
 
 - Fixed missing `failed` field from `main.TestsResults` in